+++ src-pwa/register-service-worker.js
@@ -28,7 +28,7 @@ register(process.env.SERVICE_WORKER_FILE, {
                                                     <i aria-hidden="true" class="q-icon material-icons">warning</i>
                                                   </div>
                                                   <div class="q-alert-content col self-center">
-                                                    <div>The new version of MQQT Board is available. Refresh the page to update?</div>
+                                                    <div>The new version of MQTT Board is available. Refresh the page to update?</div>
                                                   </div>
                                                   <div class="q-alert-actions col-auto gutter-xs column flex-center">
                                                     <div class="full-width">
@@ -53,7 +53,21 @@ register(process.env.SERVICE_WORKER_FILE, {
                                             </div>`
     let buttons = notification.getElementsByTagName('button'),
       body = document.getElementsByTagName('body')[0]
-    buttons[0].addEventListener('click', (ev) => { window.location.reload() })
+    function reload () {
+      if ('serviceWorker' in navigator) {
+        navigator.serviceWorker.getRegistration().then((reg) => {
+          if (reg) {
+            reg.unregister().then(() => { window.location.reload(true) })
+          } else {
+            window.location.reload(true)
+          }
+        })
+      } else {
+        window.location.reload(true)
+      }
+      setTimeout(() => { window.location.reload(true) }, 1000)
+    }
+    buttons[0].addEventListener('click', (ev) => { reload() })
     buttons[1].addEventListener('click', (ev) => { notification.remove() })
     body.appendChild(notification)
   },